Implement ctrl-p and ctrl-n to cycle through history
Ctrl-p and Ctrl-n can now be used to cycle through the history like the
default bindings in fish shell itself (and bash).

If a search was already performed, the cycling will start at that
history entry.

The default PROMPT macro was modified to support a different
visualization when cycling through history. No direction and no search
index will be printed while cycling.