Correct sanitation of newlines
Fish stores newline characters as strings in its history file.
This commit corrects the handling of such newline characters to convert
them to real newline characters again.

If the newline character is escaped with a backslash it will still be
treated as a string and only the escaping backslash is removed (as was
already the case).

For the example the previous implementation would convert this:

  for f in *.pdf
    echo -e "\t$f\n"
  end

to this:

  for f in *.pdf\necho -e "\t$f\n"\nend

This commit corrects this and converts it to this:

  for f in *.pdf
  echo -e "\t$f\n"
  end

No other control characters, apart from newline, are handled specially.
But I am also not aware of any way of inserting such literal control
characters in fish.