Support specifying start index via environment variable
The environment variable `START_INDEX` is used to specify the start
index for a scroll event.

This allows binding a key to start scrolling the history backwards even
if there is some content in the current commandline (a case which is not
supported in fish itself). The rationale for doing so is described in
https://github.com/fish-shell/fish-shell/issues/1671 and
https://github.com/fish-shell/fish-shell/issues/1668.

A fish function is provided for binding that behaviour to a key (like
Ctrl-p or the Up arrow).